Add logs to `crate::deriv`
`crate::deriv::is_derivation` prints a bunch of logs as it checks a
derivation. The logs sort of reflect stack depth. The `logwrap`
proc-macro crate is there to provide a convenient wrapper for functions
in `cands` that increment and decrement the log stack depth counter
whenever we enter or exit a function.