Fix derive-by-Agree; add Chomsky (2001, 4bii)
Now `unwind_and_agree` replaces all occurrences of old goals with the new goal.
This is necessary because otherwise different copies of the goal will not count
as identical.

Also, `derive_by_agree` allows not just the goal to move to [Spec; probe]; it
also allows a pied-piped material to move there.  For now, anything that is
equal to or contains the goal count as pied-piped material.

Add (4bii) from Chomsky (2001).