Add theory12 tests; fix wh features labels and Cyclic-Transfer
Cyclic-Transfer used to call Transfer(Comp, Comp), but it should be
calling Transfer(Phase, Comp).  This meant it used to spell-out a lower
copy of a wh-operator moved to [Spec; CP].  Fix that.

"wh" used to be a trigger feature, and this is checked by merging two
things with "wh" on each SO.  But you can't merge two SOs that each has
some trigger feature.  Separate "wh" into "uwh" vs. "iwh"; "uwh" is a
trigger feature.