Fix Theory 1
We used to test if for any X, Y in^* "some [Spec; TP]" in^* the same
root, whether their numbers of occurrences in the root are equal.  This
rules out too many cases, consider:

    We expect them to be believed to appear.

Not sure if this is grammatical or not, but this derivation should not
be ruled out by our constraint.  But if we let X = 'we' and Y = 'them',
then X occurs twice while Y occurs three times, so this derivation will
not pass the check.  The point here is that we should only compare X
and Y if they are in^* the same [Spec; TP].  This is our fix.